Ajustes na função de deletar e adicionando comentarios

+++ app/models/aluno.py
@@ -2,6 +2,7 @@ from app import db
 from flask import jsonify, make_response
 
 class Aluno(db.Model):
+    # Configuração do nome e colunas da tabela
     __tablename__ = 'aluno'
     cpf = db.Column(db.Integer, primary_key=True)
     nome = db.Column(db.String(255))
@@ -22,6 +23,7 @@ class Aluno(db.Model):
         self.AV2 = AV2
         self.media = media
 
+    # Retorna um json do model
     def json(self):
         return {
             'cpf': self.cpf,
@@ -34,15 +36,19 @@ class Aluno(db.Model):
             'media': self.media
         }
     
+    # Função para adicionar aluno no banco
     def salvar_aluno(self):
         try:
             db.session.add(self)
+            # Salva as alterações no banco
             db.session.commit()
         except Exception as e:
             print(e)
-    
+
+    # Função para alterar aluno no banco
     def modificar_aluno(self, cpf, nome, data_nascimento, sexo, idade, AV1, AV2, media):
         try:
+            # Busca o aluno pelo cpf, se ele existir atualiza e retorna TRUE, senão ele retorna FALSE
             alunoSelecionado = [alunos.json() for alunos in db.session.query(Aluno).filter(Aluno.cpf==cpf).all()]
             if(len(alunoSelecionado) == 1):
                 db.session.query(Aluno).filter(Aluno.cpf==cpf).update({'cpf': cpf,
@@ -53,16 +59,25 @@ class Aluno(db.Model):
                     'AV1': AV1,
                     'AV2': AV2,
                     'media': media})
-                db.session.commit()
+                # Salva as alterações no banco
+                db.session.commit() 
                 return True
             else:
                 return False
         except Exception as e:
             print(e)
     
+    # Função para deletar aluno no banco
     def deletar_aluno(self, cpf):
         try:
-            db.session.query(Aluno).filter(Aluno.cpf==cpf).delete()
-            db.session.commit()
+            # Busca o aluno pelo cpf, se ele existir deleta e retorna TRUE, senão ele retorna FALSE
+            alunoSelecionado = [alunos.json() for alunos in db.session.query(Aluno).filter(Aluno.cpf==cpf).all()]
+            if(len(alunoSelecionado) == 1):
+                db.session.query(Aluno).filter(Aluno.cpf==cpf).delete()
+                # Salva as alterações no banco
+                db.session.commit() 
+                return True
+            else:
+                return False
         except Exception as e:
             print(e)
